EAMxx: split field.hpp into decl and impl files
