+++ components/eamxx/src/share/field/field.hpp
@@ -1,18 +1,14 @@
 #ifndef SCREAM_FIELD_HPP
 #define SCREAM_FIELD_HPP
 
-#include "ekat/std_meta/ekat_std_type_traits.hpp"
 #include "share/field/field_header.hpp"
 #include "share/scream_types.hpp"
 
-#include "ekat/ekat_type_traits.hpp"
-#include "ekat/kokkos/ekat_kokkos_meta.hpp"
-#include "ekat/kokkos/ekat_kokkos_utils.hpp"
+#include "ekat/std_meta/ekat_std_type_traits.hpp"
 #include "ekat/kokkos/ekat_subview_utils.hpp"
 
 #include <memory>   // For std::shared_ptr
 #include <string>
-#include <type_traits>
 
 namespace scream
 {
@@ -251,10 +247,10 @@ public:
 
 protected:
 
-  template<typename ST, HostOrDevice HD = Device>
+  template<HostOrDevice HD, typename ST>
   void deep_copy_impl (const ST value);
 
-  template<typename ST, HostOrDevice HD = Device>
+  template<HostOrDevice HD, typename ST>
   void deep_copy_impl (const Field& field_src);
 
   template<HostOrDevice HD>
@@ -304,378 +300,9 @@ inline bool operator== (const Field& lhs, const Field& rhs) {
   return lhs.get_header().get_identifier() == rhs.get_header().get_identifier();
 }
 
-// ================================= IMPLEMENTATION ================================== //
-
-template<typename DT, HostOrDevice HD>
-auto Field::get_view () const
- -> get_view_type<DT,HD>
-{
-  // The destination view type on correct mem space
-  using DstView = get_view_type<DT,HD>;
-  // The dst value types
-  using DstValueType = typename DstView::traits::value_type;
-  // The ViewDimension object from the Dst View (used to check validity of possible compile-time extents)
-  using dims_type = typename DstView::traits::dimension;
-  // We only allow to reshape to a view of the correct rank
-  constexpr int DstRank = DstView::rank;
-  constexpr int DstRankDynamic= DstView::rank_dynamic;
-
-  // Make sure input field is allocated
-  EKAT_REQUIRE_MSG(is_allocated(),
-      "Error! Cannot extract a field's view before allocation happens.\n");
-
-  EKAT_REQUIRE_MSG (not m_is_read_only || std::is_const<DstValueType>::value,
-      "Error! Cannot get a view to non-const data if the field is read-only.\n");
-
-  // Get src details
-  const auto& alloc_prop = m_header->get_alloc_properties();
-  const auto& field_layout = m_header->get_identifier().get_layout();
-
-  EKAT_REQUIRE_MSG(DstRank==field_layout.rank(),
-      "Error! You can only reshape to a view of the correct rank (equal to the FieldLayout's one).\n");
-
-  // Check the reinterpret cast makes sense for the Dst value types (need integer sizes ratio)
-  EKAT_REQUIRE_MSG(alloc_prop.template is_compatible<DstValueType>(),
-      "Error! Source field allocation is not compatible with the requested value type.\n");
-
-  // Start by reshaping into a ND view with all dynamic extents
-  const auto view_ND = get_ND_view<HD,DstValueType,DstRank>();
-
-  using dyn_DT = typename decltype(view_ND)::traits::data_type;
-  if (!std::is_same<dyn_DT,DT>::value) {
-    // The user requested some compile-time dimensions.
-    // Let's check that they are correct
-    for (int i=DstRankDynamic; i<DstRank; ++i) {
-      EKAT_REQUIRE_MSG(view_ND.extent(i)==dims_type::static_extent(i),
-          "Error! The template DataType contains an invalid compile-time dimensions:\n"
-          "    - field name: " + m_header->get_identifier().name() + "\n"
-          "    - dim index: " + std::to_string(i) + "\n"
-          "    - input compile time dimension: " + std::to_string(dims_type::static_extent(i)) + "\n"
-          "    - field internal dimension: " + std::to_string(view_ND.extent(i)) + "\n");
-    }
-  }
-
-  // Before building the DstView from view_ND, we have one more check:
-  // if DstRankDynamic==0, kokkos specializes the view offset struct,
-  // assuming *no* stride. That's fine, as long as this field alloc
-  // props ensure that there is no stride
-  EKAT_REQUIRE_MSG (DstRankDynamic>0 || alloc_prop.contiguous(),
-      "Error! Cannot use all compile-time dimensions for strided views.\n");
-
-  return DstView(view_ND);
-}
-
-template<typename DT, HostOrDevice HD>
-auto Field::get_strided_view () const
- -> get_strided_view_type<DT,HD>
-{
-  // The destination view type on correct mem space
-  using DstView = get_strided_view_type<DT,HD>;
-  // The dst value types
-  using DstValueType = typename DstView::traits::value_type;
-  // We only allow to reshape to a view of the correct rank
-  constexpr int DstRank = DstView::rank;
-  constexpr int DstRankDynamic= DstView::rank_dynamic;
-
-  // Get src details
-  const auto& alloc_prop = m_header->get_alloc_properties();
-  const auto& fl = m_header->get_identifier().get_layout();
-
-  // Checks
-  EKAT_REQUIRE_MSG (DstRank==1 && fl.rank()==1,
-      "Error! Strided view only available for rank-1 fields.\n");
-  EKAT_REQUIRE_MSG (DstRankDynamic==1,
-      "Error! Strided view not allowed with compile-time dimensions.\n");
-  EKAT_REQUIRE_MSG(is_allocated(),
-      "Error! Cannot extract a field's view before allocation happens.\n");
-  EKAT_REQUIRE_MSG (not m_is_read_only || std::is_const<DstValueType>::value,
-      "Error! Cannot get a view to non-const data if the field is read-only.\n");
-  EKAT_REQUIRE_MSG(alloc_prop.template is_compatible<DstValueType>(),
-      "Error! Source field allocation is not compatible with the requested value type.\n");
-
-  // Check if this field is a subview of another field
-  const auto parent = m_header->get_parent().lock();
-  if (parent!=nullptr) {
-    // Parent field has correct layout to reinterpret the view into N+1-dim view
-    // So create the parent field on the fly, use it to get the N+1-dim view, then subview it.
-    // NOTE: we can set protected members, since f is the same type of this class.
-    Field f;
-    f.m_header = parent;
-    f.m_data   = m_data;
-
-    // Take 2 dimensional view with normal LayoutRight
-    auto v_np1 = f.get_ND_view<HD,DstValueType,2>();
-
-    // Now we can subview v_np1 at the correct slice
-    const auto& info = m_header->get_alloc_properties().get_subview_info();
-    const int idim = info.dim_idx;
-    const int k    = info.slice_idx;
-
-    // So far we can only subview at first or second dimension.
-    EKAT_REQUIRE_MSG (idim==0 || idim==1,
-        "Error! Subview dimension index is out of bounds.\n");
-
-    // Use correct subview utility
-    if (idim==0) {
-      return DstView(ekat::subview(v_np1,k));
-    } else {
-      return DstView(ekat::subview_1(v_np1,k));
-    }
-  }
-
-  // Not a subfield, so stride=1, and we can create the strided view from the LayoutRight 1d view.
-  return DstView(get_ND_view<HD,DstValueType,1>());
-}
-
-template<HostOrDevice HD>
-void Field::
-deep_copy (const Field& field_src) {
-  EKAT_REQUIRE_MSG (not m_is_read_only,
-      "Error! Cannot call deep_copy on read-only fields.\n");
-
-  EKAT_REQUIRE_MSG (data_type()==field_src.data_type(),
-      "Error! Cannot copy fields with different data type.\n");
-
-  switch (data_type()) {
-    case DataType::IntType:
-      deep_copy_impl<int,HD>(field_src);
-      break;
-    case DataType::FloatType:
-      deep_copy_impl<float,HD>(field_src);
-      break;
-    case DataType::DoubleType:
-      deep_copy_impl<double,HD>(field_src);
-      break;
-    default:
-      EKAT_ERROR_MSG ("Error! Unrecognized field data type in Field::deep_copy.\n");
-  }
-}
-
-template<typename ST, HostOrDevice HD>
-void Field::
-deep_copy (const ST value) {
-  EKAT_REQUIRE_MSG (not m_is_read_only,
-      "Error! Cannot call deep_copy on read-only fields.\n");
-
-  const auto my_data_type = data_type();
-  switch (my_data_type) {
-    case DataType::IntType:
-      EKAT_REQUIRE_MSG( (std::is_convertible<ST,int>::value),
-          "Error! Input value type is not convertible to field data type.\n"
-          "   - Input value type: " + ekat::ScalarTraits<ST>::name() + "\n"
-          "   - Field data type : " + e2str(my_data_type) + "\n");
-      deep_copy_impl<int,HD>(value);
-      break;
-    case DataType::FloatType:
-      EKAT_REQUIRE_MSG( (std::is_convertible<ST,float>::value),
-          "Error! Input value type is not convertible to field data type.\n"
-          "   - Input value type: " + ekat::ScalarTraits<ST>::name() + "\n"
-          "   - Field data type : " + e2str(my_data_type) + "\n");
-      deep_copy_impl<float,HD>(value);
-      break;
-    case DataType::DoubleType:
-      EKAT_REQUIRE_MSG( (std::is_convertible<ST,double>::value),
-          "Error! Input value type is not convertible to field data type.\n"
-          "   - Input value type: " + ekat::ScalarTraits<ST>::name() + "\n"
-          "   - Field data type : " + e2str(my_data_type) + "\n");
-      deep_copy_impl<double,HD>(value);
-      break;
-    default:
-      EKAT_ERROR_MSG ("Error! Unrecognized field data type in Field::deep_copy.\n");
-  }
-}
-
-template<typename ST, HostOrDevice HD>
-void Field::
-deep_copy_impl (const Field& field_src) {
-
-  const auto& layout     = get_header().get_identifier().get_layout();
-  const auto& layout_src = field_src.get_header().get_identifier().get_layout();
-  EKAT_REQUIRE_MSG(layout==layout_src,
-       "ERROR: Unable to copy field " + field_src.get_header().get_identifier().name() + 
-          " to field " + get_header().get_identifier().name() + ".  Layouts don't match.");
-  const auto  rank = layout.rank();
-  // Note: we can't just do a deep copy on get_view_impl<HD>(), since this
-  //       field might be a subfield of another. We need the reshaped view.
-  switch (rank) {
-    case 1:
-      {
-        auto v     = get_view<ST*,HD>();
-        auto v_src = field_src.get_view<const ST*,HD>();
-        Kokkos::deep_copy(v,v_src);
-      }
-      break;
-    case 2:
-      {
-        auto v     = get_view<ST**,HD>();
-        auto v_src = field_src.get_view<const ST**,HD>();
-        Kokkos::deep_copy(v,v_src);
-      }
-      break;
-    case 3:
-      {
-        auto v     = get_view<ST***,HD>();
-        auto v_src = field_src.get_view<const ST***,HD>();
-        Kokkos::deep_copy(v,v_src);
-      }
-      break;
-    case 4:
-      {
-        auto v     = get_view<ST****,HD>();
-        auto v_src = field_src.get_view<const ST****,HD>();
-        Kokkos::deep_copy(v,v_src);
-      }
-      break;
-    case 5:
-      {
-        auto v     = get_view<ST*****,HD>();
-        auto v_src = field_src.get_view<const ST*****,HD>();
-        Kokkos::deep_copy(v,v_src);
-      }
-      break;
-    default:
-      EKAT_ERROR_MSG ("Error! Unsupported field rank in 'deep_copy'.\n");
-  }
-}
-
-template<typename ST, HostOrDevice HD>
-void Field::deep_copy_impl (const ST value) {
-
-  // Note: we can't just do a deep copy on get_view_impl<HD>(), since this
-  //       field might be a subfield of another. Instead, get the
-  //       reshaped view first, based on the field rank.
-
-  const auto& layout = get_header().get_identifier().get_layout();
-  const auto  rank   = layout.rank();
-  switch (rank) {
-    case 1:
-      {
-        auto v = get_view<ST*,HD>();
-        Kokkos::deep_copy(v,value);
-      }
-      break;
-    case 2:
-      {
-        auto v = get_view<ST**,HD>();
-        Kokkos::deep_copy(v,value);
-      }
-      break;
-    case 3:
-      {
-        auto v = get_view<ST***,HD>();
-        Kokkos::deep_copy(v,value);
-      }
-      break;
-    case 4:
-      {
-        auto v = get_view<ST****,HD>();
-        Kokkos::deep_copy(v,value);
-      }
-      break;
-    case 5:
-      {
-        auto v = get_view<ST*****,HD>();
-        Kokkos::deep_copy(v,value);
-      }
-      break;
-    case 6:
-      {
-        auto v = get_view<ST******,HD>();
-        Kokkos::deep_copy(v,value);
-      }
-      break;
-    default:
-      EKAT_ERROR_MSG ("Error! Unsupported field rank in 'deep_copy'.\n");
-  }
-}
-
-template<HostOrDevice HD,typename T,int N>
-auto Field::get_ND_view () const ->
-  if_t<(N<MaxRank),get_view_type<data_nd_t<T,N>,HD>>
-{
-  const auto& fl = m_header->get_identifier().get_layout();
-  EKAT_REQUIRE_MSG (N==1 || N==fl.rank(),
-      "Error! Input Rank must either be 1 (flat array) or the actual field rank.\n");
-
-  // Check if this field is a subview of another field
-  const auto parent = m_header->get_parent().lock();
-  if (parent!=nullptr) {
-    // Parent field has correct layout to reinterpret the view into N+1-dim view
-    // So create the parent field on the fly, use it to get the N+1-dim view, then subview it.
-    // NOTE: we can set protected members, since f is the same type of this class.
-    Field f;
-    f.m_header = parent;
-    f.m_data   = m_data;
-
-    auto v_np1 = f.get_ND_view<HD,T,N+1>();
-
-    // Now we can subview v_np1 at the correct slice
-    const auto& info = m_header->get_alloc_properties().get_subview_info();
-    const int idim = info.dim_idx;
-    const int k    = info.slice_idx;
-
-    // So far we can only subview at first or second dimension.
-    EKAT_REQUIRE_MSG (idim==0 || idim==1,
-        "Error! Subview dimension index is out of bounds.\n");
-
-    EKAT_REQUIRE_MSG (idim==0 || N>1,
-        "Error! Cannot subview a rank-2 (or less) view along 2nd dimension without losing LayoutRight.\n");
-
-    // Use SFINAE-ed get_subview helper function to pick correct
-    // subview impl. If N+1<=2 and idim!=0, the code craps out in the check above.
-    if (idim==0) {
-      return ekat::subview(v_np1,k);
-    } else {
-      return get_subview_1<HD,T,N+1>(v_np1,k);
-    }
-  }
-
-  // Compute extents from FieldLayout
-  const auto& alloc_prop = m_header->get_alloc_properties();
-  auto num_values = alloc_prop.get_alloc_size() / sizeof(T);
-  Kokkos::LayoutRight kl;
-  for (int i=0; i<N; ++i) {
-    if (i==N-1) {
-      kl.dimension[i] = num_values;
-    } else {
-      kl.dimension[i] = fl.dim(i);
-      num_values = fl.dim(i)==0 ? 0 : num_values/fl.dim(i);
-    }
-  }
-  auto ptr = reinterpret_cast<T*>(get_view_impl<HD>().data());
-
-  using ret_type = get_view_type<data_nd_t<T,N>,HD>;
-
-  return ret_type (ptr,kl);
-}
-
-template<HostOrDevice HD,typename T,int N>
-auto Field::get_ND_view () const ->
-  if_t<N==MaxRank,get_view_type<data_nd_t<T,N>,HD>>
-{
-  const auto& fl = m_header->get_identifier().get_layout();
-  EKAT_REQUIRE_MSG (N==1 || N==fl.rank(),
-      "Error! Input Rank must either be 1 (flat array) or the actual field rank.\n");
-
-  // Given that N==MaxRank, this field cannot be a subview of another field
-  EKAT_REQUIRE_MSG (m_header->get_parent().expired(),
-      "Error! A view of rank " + std::to_string(MaxRank) + " should not be the subview of another field.\n");
-
-  // Compute extents from FieldLayout
-  const auto& alloc_prop = m_header->get_alloc_properties();
-  auto num_values = alloc_prop.get_alloc_size() / sizeof(T);
-  Kokkos::LayoutRight kl;
-  for (int i=0; i<N-1; ++i) {
-    kl.dimension[i] = fl.dim(i);
-    num_values /= fl.dim(i);
-  }
-  kl.dimension[N-1] = num_values;
-  auto ptr = reinterpret_cast<T*>(get_view_impl<HD>().data());
-
-  using ret_type = get_view_type<data_nd_t<T,N>,HD>;
-  return ret_type (ptr,kl);
-}
-
 } // namespace scream
+
+// Include template methods implementation
+#include "share/field/field_impl.hpp"
 
 #endif // SCREAM_FIELD_HPP
